save and load preferences, exceptions, key bindings
Added the preferences file which is automatically created if it isnt found in the current working directory. Program loads in the settings from the file if it is found. You can save your preferences to the file or manually edit it. Exceptions handling has been added. Added key bindings for easy navigation. Extra option to change font.